lambdas/uploadFn/requirements.txt
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
boto3==1.34.162
